Put the client_uid in all instances of a single locaiton.
Previously you only saw the small client snippet when on a location
detail page, since in other contexts we were listing Locations beneath a
particular client. When parsing the individual Locations, it just makes
your life easier to have the client UID available with the location's
attributes. Otherwise in some contexts you'd need to step up a level to
figure out which client the location was associated with.