Don't protect from forgery
I tried modifying the rack-cors gem as advised via the interwebs but still got
the same error. The only way I can get this to work is to NOT protect from
forgery on the index action. Since this is a GET request I think this may be ok,
but looking for opinions.